Refactor RecordsService
